+++ Pandora/Paradigm/Structure/Ability/Accessible.hs
@@ -2,5 +2,5 @@ module Pandora.Paradigm.Structure.Ability.Accessible where
 
 import Pandora.Paradigm.Inventory.Optics (type (:-.))
 
-class Accessible a e where
-	access :: a :-. e
+class Accessible tgt src where
+	access :: src :-. tgt